Create teams mgr
